fix: ensure float32 audio dtype to prevent pyannote CUDA batch norm crash
sf.read() returns float64 by default. When the audio sample rate matches
the target (no resampling needed), the float64 array passed through to
pyannote's diarization pipeline, causing a DoubleTensor vs FloatTensor
mismatch on CUDA batch norm layers.

+++ src/diarized_transcriber/transcription/audio.py
@@ -37,9 +37,11 @@ def load_audio(
             # Use whisperx's built-in audio loading for URLs
             audio = whisperx.load_audio(str(source))
             return audio, sample_rate
-        
+
         # For file paths and file-like objects, use soundfile
-        audio_data, file_sample_rate = sf.read(source)
+        # dtype="float32" avoids float64 default which causes dtype
+        # mismatches in pyannote's batch norm layers on CUDA
+        audio_data, file_sample_rate = sf.read(source, dtype="float32")
         
         # Resample if necessary
         if file_sample_rate != sample_rate:
